feature/cds-139 Removed separate files reducing size of repositories
