--- datahub/bed_api/repositories/__init__.py
@@ -1,13 +1,0 @@
-from datahub.bed_api.repositories.account_repository import AccountRepository  # noqa:F401
-from datahub.bed_api.repositories.contact_repository import ContactRepository  # noqa:F401
-from datahub.bed_api.repositories.event_attendee_repository import (  # noqa:F401
-    EventAttendeeRepository,
-)
-from datahub.bed_api.repositories.event_repository import EventRepository  # noqa:F401
-from datahub.bed_api.repositories.policy_issues_repository import (  # noqa:F401
-    PolicyIssuesRepository,
-)
-from datahub.bed_api.repositories.base_repository import (  # noqa:F401
-    ReadRepository,
-    ReadWriteRepository,
-)
