--- datahub/bed_api/repositories/contact_repository.py
@@ -1,12 +1,0 @@
-from datahub.bed_api.repositories.base_repository import ReadWriteRepository
-
-
-class ContactRepository(ReadWriteRepository):
-    """
-    Contact Repository to connect to BED Salesforce API
-
-    Repository pattern for Salesforce interactions with Contacts data
-    https://loginhub--november.lightning.force.com/lightning/setup/ObjectManager/Contact/FieldsAndRelationships/view
-    """
-
-    entity_name = 'Contact'
