--- datahub/bed_api/repositories/event_attendee_repository.py
@@ -1,12 +1,0 @@
-from datahub.bed_api.repositories.base_repository import ReadWriteRepository
-
-
-class EventAttendeeRepository(ReadWriteRepository):
-    """
-    Event Attendee Repository to connect to BED Salesforce API
-
-    Repository pattern for Salesforce interactions with Event Attendee data
-    https://loginhub--november.lightning.force.com/lightning/setup/ObjectManager/01I58000001EcAX/FieldsAndRelationships/view
-    """
-
-    entity_name = 'Event_Attendee__c'
